Cleanup scrapers (#24)
* Simplify RealScraperContext

* Start simplifying regexes

* Add some helpers

* Factor out Shopify stuff

* Keep cleaning up

* Moving trimming into Executor

* Refactor

* Keep tidying

* Continue cleaning

* Eliminate BigDecimal

+++ src/jvmMain/kotlin/choliver/neapi/Model.kt
@@ -1,14 +1,13 @@
 package choliver.neapi
 
-import java.math.BigDecimal
 import java.net.URI
 
 data class ParsedItem(
   val name: String,
   val summary: String? = null,
-  val pricePerCan: BigDecimal,
+  val pricePerCan: Double,
   val sizeMl: Int? = null,
-  val abv: BigDecimal? = null,
+  val abv: Double? = null,
   val available: Boolean,
   val thumbnailUrl: URI? = null,
   val url: URI
